Rename vmware_guest_tools_upgrade to vmware_guest_tools.
Add "state" parameter.
Add logic for state, with 3 available states: upgrade, mount, and unmount.
Add code to perform VMWare Tools installer mount and unmount.
Update basic documentation.
Update upgrade example.
Add examples for mount and unmount state.
Add additional author.
